finished CRUD product app with Chakra UI and Zustand state

+++ backend/controllers/product.controller.js
@@ -34,10 +34,11 @@ export const createProduct = async (req, res) => {
 export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const product = req.body;
+
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res
       .status(404)
-      .json({ success: false, message: "Product not found" });
+      .json({ success: false, message: "Invalid Product ID" });
   }
 
   try {
@@ -53,6 +54,12 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product ID" });
+  }
 
   try {
     await Product.findByIdAndDelete(id);
